refactor(slack-app): move shared dataclasses into slack_app.types
Move the five workflow-boundary dataclasses out of the two posthog_code
workflow files and into the shared types module so the activities that
this PR series will split out can take them as typed signatures without
reaching back into the workflow files.

Moves:
- PostHogCodeRepoCascadeOutcome
- SlackRepoSelectionOutcome
- PostHogCodeRulesCommandResult
- PostHogCodeSlackMentionCommandWorkflowInputs
- PostHogCodeSlackMentionCommandResult

products/slack_app/backend/api.py updates its imports to take the
inputs dataclasses from slack_app directly instead of relying on the
workflow files re-exporting them.
