+++ posthog/temporal/ai/slack_app/__init__.py
@@ -1,9 +1,10 @@
 """Temporal activities for the PostHog Slack App.
 
 Activities living here run on ``settings.MAX_AI_TASK_QUEUE`` and are imported
-by ``posthog/temporal/ai/posthog_code_slack_mention.py`` at module load time.
-The package is the long-term home for activity splits out of that file; this
-PR ships the classifier and follow-ups will land here one PR at a time.
+by ``posthog/temporal/ai/posthog_code_slack_mention.py`` and
+``posthog/temporal/ai/posthog_code_slack_mention_command.py`` at module load
+time. The package is the long-term home for activity splits out of those
+files; activities land here one PR at a time.
 """
 
 from posthog.temporal.ai.slack_app.activities import (
@@ -11,7 +12,14 @@ from posthog.temporal.ai.slack_app.activities import (
     classify_message_is_agent_directed,
     classify_untagged_followup_activity,
 )
-from posthog.temporal.ai.slack_app.types import PostHogCodeSlackMentionWorkflowInputs
+from posthog.temporal.ai.slack_app.types import (
+    PostHogCodeRepoCascadeOutcome,
+    PostHogCodeRulesCommandResult,
+    PostHogCodeSlackMentionCommandResult,
+    PostHogCodeSlackMentionCommandWorkflowInputs,
+    PostHogCodeSlackMentionWorkflowInputs,
+    SlackRepoSelectionOutcome,
+)
 
 SLACK_APP_ACTIVITIES = [
     classify_untagged_followup_activity,
@@ -19,8 +27,13 @@ SLACK_APP_ACTIVITIES = [
 
 __all__ = [
     "CLASSIFIER_THREAD_HISTORY_MESSAGES",
+    "PostHogCodeRepoCascadeOutcome",
+    "PostHogCodeRulesCommandResult",
+    "PostHogCodeSlackMentionCommandResult",
+    "PostHogCodeSlackMentionCommandWorkflowInputs",
     "PostHogCodeSlackMentionWorkflowInputs",
     "SLACK_APP_ACTIVITIES",
+    "SlackRepoSelectionOutcome",
     "classify_message_is_agent_directed",
     "classify_untagged_followup_activity",
 ]
